+++ htdocs/_PX/_FW/pxcommands/config.php
@@ -126,7 +126,9 @@ class px_pxcommands_config extends px_bases_pxcommand{
 		$src .= $this->mk_config_unit('system.default_theme_id','デフォルトのテーマID');
 		$src .= $this->mk_config_unit('system.filesystem_encoding','ファイル名の文字エンコード');
 		$src .= $this->mk_config_unit('system.output_encoding','出力エンコード');
-		$src .= $this->mk_config_unit('system.output_eof_coding',' 出力改行コード("CR"|"LF"|"CRLF")');
+		$src .= $this->mk_config_unit('system.output_eof_coding','出力改行コード("CR"|"LF"|"CRLF")');
+		$src .= $this->mk_config_unit('system.file_default_permission','ファイル書き込み時のデフォルトのパーミッション');
+		$src .= $this->mk_config_unit('system.dir_default_permission','ディレクトリ書き込み時のデフォルトのパーミッション');
 		$src .= '</table>' . "\n";
 
 		$src .= '</div><!-- /.unit -->'."\n";